chore: moved census client event definitions
